Don't depend on a hardcoded java ls jar version
The java language server has no stable release. However, there is no
reason to depend on an arbitrary snapshot of the java language
server. The version is a leftover of the initial pull request adding
support for the java language server, which installed the server as
well, so we can remove it.