Bring in lsp-project-whitelist
As an alternative to 'lsp-project-whitelist', this allows listing the projects
for which to use LSP mode.

The current logic is that if the whitelist is not nil, then the blacklist is
ignored.  This can be changed if a different policy makes sense, the decision
process is broken out into 'lsp--should-start-p'