Make fallback values for sortText and insertText match the spec
Prior to this change, if a language server returned a falsy value for
sortText or insertText (such as the empty string), lsp-mode would throw
an error trying to process it, since gethash returns nil in those cases.

By introducing a wrapper around gethash that uses the default when any
falsey value is returned, not just when the passed key does not exist.