Avoid deleting overlays created by Flycheck (fix #87)
Previously we saved only the positions of the overlays created for highlighting
symbols, and called `remove-overlays` with this position.  This had the
side-effect of deleting *any* overlay within this range, not just the ones
created by lsp-mode for highlighting.

This commit saves the overlays created when highlighting directly in a list, and
deletes them using `delete-overlay`.