Allow the mode-line to change dynamically
This allows us to give some feedback to the user when the language server is
busy.

E.g., the RLS sends a `rustDocument/beginBuild` message when it starts building
the workspace, which can take quite some time on larger projects.

Since this is a new API, the version string is increased accordingly.