Split the lsp-XXXX-enable into non-interactive and interactive
The non-interactive one gets called in the mode hook, and is quiet

The interactive one will ask to add a project to the whitelist if invoked to
start LSP for it and it would not otherwise start, if `lsp-enable-whitelist-ask'
is set.

lsp-MAJOR-MODE-enable now always starts the project

Regardless of whitelist/blacklist settings.

And the whitelist itself is manipulated via

  lsp-MAJOR-MODE-whitelist-add
  lsp-MAJOR-MODE-whitelist-remove