Allow other packages to register client capabilities (#157)
Adds two functions for registering/unregistering client capabilities:
`lsp-register-client-capabilities` and `lsp-unregister-client-capabilities`.

Registered capabilities are stored in the `extra-client-capabilities` slot in
the current workspace. The `extra-client-capabilitites` slot is an alist in the
format of `(PACKAGE-NAME . CAPS)` cons cell. CAPS can be either a plist, nil, or
a function.

All registerd client capabilities are merged into the default client
capabilities. The `lsp--merge-two-plists` function implements a deep merging
algorithm for plists. Existing leaf capability entry cannot be overwritten.