Fixes bug with default value of `language-fn-id`
cl-defmacro automatically evaluates default values. To prevent evaluation, the
default value needs to be escaped, but when escaped `language-id` is not
replaced by the macro system. To circumvent these problems, the value of
`language-fn-id` is checked at execution and if the value is `nil` a lambda
returning the `language-id` is used.