Support initializeOptions when starting an LSP server
LSP v3.0 spec defines an optional member of initialize:InitializeParams,
initializeOptions, that can be used to pass user-provided initialization
options to the server.

Implements an (optional) extra arg to lsp--start, extra-init-params.
This can be either a fixed plist of options or a function that takes
an LSP workspace and generates the initializeOptions params.

Existing clients should not be affected. Fixes #180