Allow clients to customize how to get prefix bounds.
Add a new :prefix-function parameter to the define functions.

Currently lsp-mode graps the symbol before point as the completion prefix.
The prefix is used for matching completion candidates and will be replaced by
the selected completion candidate. Some servers use different methods to
calculate prefixes. If the prefixes between the server and the client don't
match possible results are:

 * Candidates are filtered out unexpecedly, because the client side prefix has
   extra leading characters.
 * Completed candidate has duplicated leading characters, because the server
   side prefix has extra leading characters and all the candidates also have
   the leading characters.

Only the clients knows which server they are talking to, so it's the clients'
responsibility to implement custom prefix funtions if necessary.