Implemented multiple folders support
Fixes https://github.com/emacs-lsp/lsp-mode/issues/345

* Changed the startup code of the lsp-workspace. Now it will check whether there
is file in the workspace root which contains workspace folders. If there is such
file it will be used as a source for workspace folders. If there is no such file
it will work as before that change.
* Added methods for adding/removing folders
* Tested with JDT server which seems to be the only server supporting the
feature.
* Verified that client which does not support workspace folders work as
expected(verfied with lsp-python)