Refactor lsp-hover
The original code cancelled pending hover request, but this behavior is not used in any other language client I know about. textDocumenthover is fast so this is not very necessary.

To reduce hover requests, we cache "contents" and "range", and reuse "contents" if (point) is still within "range".

lsp--hover-callback is better using eldoc-message but after a movement operation, the echo area is cleared and eldoc-message does not redisplay the message. So use message for now.