Provide lsp-mode in correct file
The tools used to maintain the Emacsmirror automatically determine the
libraries provided by a package by looking for emacs-lisp files that
provide a feature matching their filename (sans extension obviously).
This is done to avoid false-positives.

These tools get confused if features are provided in files that do not
match, as was the case here.