Overhaul eldoc
We request both hover and signatureHelp in the eldoc callback and we
face a choice when both are available.

Add a variable lsp-eldoc-prefer-signature-help to customize the
preference (defaults to t). This variable can be extended in the future
to take triggerCharacters and evil-insert-state into account.

There are some cons to send the signatureHelp request unconditionally because:

1. signatureHelp is a feature tightly connected with completion, they
   are expensive in some language servers, and there is no good caching
   mechanism (SignatureHelp does not include a range). We may consider
   integrating triggerCharacters in the future.
2. signatureHelp may be state-aware: it is mostly useful in insert state.
   In normal state (read-only browsing), its information can largely be
   compensated by requesting hover on its function identifier.  There is a
   very little loss by preferring hover.  On the other hand, the loss of
   hover is hardly remedied by other methods.