Use bingo as go LSP server
As
sourcegraph/go-langserver@3c8059f
notes, the go-langserver is not longer maintained and additional fixes are not
added to it, so the new features (such as go mod) are not (and will not) be
supported.

The authors recommend https://github.com/saibing/bingo as alternative and my
tests show it fixes the issues I was hitting with the go-langserver for me (such
as missing flycheck).

The go-langserver is kept registered for compatibility reasons.