Change tramp implementation to not rely on TCP connections
Opening TCP ports is not very secure, and often not possible. Instead,
tunnel client invocations directly through tramp.

Since LSP is using carriage returns and does not terminate the payload
with newlines, it is necessary to wrap the remote commands with stty
settings to preserve carriage returns.

We also need to append extra newlines to the payload to work around
terminal line buffering through tramp. This is legal in JSON, and
should not cause issues.