Fix action support for servers that return a CodeAction instead of a Command
The new TypeScript server can return a list of CodeAction structures
instead of a list of Commands. This means that, with the current code,
the subsequent call to 'workspace/executeCommand' was performed with
incorrect arguments, causing the refactor to fail.

With this patch, the response of a 'textDocument/codeAction' is
correctly parsed either as a CodeAction or a Command, and tsserver,
and potentially other language servers, now perform the refactor
operation correctly.