Only update flymake diagnsotics on publishDiagnostics
This changes the way the flymake backend works a bit.

Previously, diagnostics were only updated from the flymake backend function,
which gets called whenever flymake thinks it's a good idea. This doesn't
necessarily correspond to the time when they should be updated, i.e. when new
diagnostics arrive.

Now, the flymake backend only saves the report function, which is then called
from `lsp--flymake-update-diagnostics` which is triggered on
`textDocument/publishDiagnostics`.