Show flymake diagnostics immediately when they arrive (#684)
* Only update flymake diagnsotics on publishDiagnostics

This changes the way the flymake backend works a bit.

Previously, diagnostics were only updated from the flymake backend function,
which gets called whenever flymake thinks it's a good idea. This doesn't
necessarily correspond to the time when they should be updated, i.e. when new
diagnostics arrive.

Now, the flymake backend only saves the report function, which is then called
from `lsp--flymake-update-diagnostics` which is triggered on
`textDocument/publishDiagnostics`.

* Update flymake diagnostics on first backend invokation

When we get diagnostics from the server before flymake has called our backend
function, we don't have a way to report them at that time.

With this change, the unreported diagnostics are now passed to flymake as soon
as possible, i.e when it first calls our backend function.