Fix return value of lsp-eldoc-function
The documentation of `eldoc-documentation-function` says:

> It should return nil if there's no doc appropriate for the context.

`lsp-eldoc-function` was always returning eldoc-last-message. Because
of this, the last message displayed by eldoc would not disappear when
the point is moved in a place where no doc is available or
lsp-eldoc-hook returns nil and doesn't call lsp-hover.

The downside is that moving the point inside the same identifier might
trigger a refresh of the eldoc message, even if it is the same. Maybe
it is possible to fix this problem by changing `lsp-hover`. So that it
returns the documentation string instead of calling `eldoc-message`.