Fix warnings produced by lsp-fsharp (#889)
While testing the lsp-fsharp, I was getting multiple warnings. Seems
like there was wrong usage of `if` instead of `unless` + couple of
notifications produced by the fsautocomplete that were not configured
and caused the warnings to produced.

I think marking the notifications to be ignored until we have some use
for them is better then showing the annoying warning messages.