Add new lsp-deferred entry point.
You can use lsp-deferred instead of lsp in your major mode hooks. This
will defer starting the LSP server and adding the file to the LSP
server until the buffer is actually visible in Emacs. This can make
startup a lot faster because it avoids sending all open files at once
to the LSP server when starting Emacs or restarting the workspace.

Fixes #792