lsp-go: support adding modules to library folders
Recent versions of go added the concept of modules as a first class
system for managing third party dependencies. Modules are installed to
$GOPATH/pkg/mod and are set to have read-only file permissions. The
same module+version will be shared across multiple projects.

Jumping to these files should not trigger a prompt for starting a new
lsp server.

This adds logic to detect the modules directory and adds it to
library-folders automatically. This behavior can be disabled with the
customization `lsp-clients-go-library-directories-include-go-modules'.

We invoke the go tool to determine the path of the modules directory
instead of looking at the GOPATH environment variable directly. This
is necessary to determine the default GOPATH directory when the
environment variable isn't set.