Fix go lib default dirs to handle remote buffers
Use process-file which is TRAMP aware and will call process on the
remote machine if the default-directory is a remote directory.

If we are in a remote buffer we append the remote connection identity
prefix to each path. Without this we fail to identify a remote file as
being in a library directory.