Fix hover text not updating
When moving point from one position with one hover text directly to
another position with a different hover text, the echo area was not
being updated; instead, the previous hover text was being redisplayed.

This was due to the previous message being saved in
lsp--eldoc-saved-message, and redisplayed from there. The only
circumstance in which the variable was cleared was when point was
moved to whitespace and "hovered" there.

Fix this by clearing lsp--eldoc-saved-message as soon as it is clear
that its contents is no longer valid, i.e., before sending a
textDocument/hover request.