Fix handling textDocument/prepareRename returning an object
The current version of the specification documents that
textDocument/prepareRename may return
`Range | { range: Range, placeholder: string } | null`.

Currently, the second case is not handled, and should a server return
such an object, it causes a "Wrong type argument: hash-table-p, nil"
error in lsp-mode.

Fix this by checking if the returned value is a hash containing a
"range" key, and if so, process it instead of the return value.

Handling of the placeholder field is not implemented in this commit.