Apply documentChanges in order
With the introduction of RenameFile operations in
WorkspaceEdit.documentChanges, the server may send sequences such as:

- Perform edits on a file, then rename it
- Rename a file, then perform edits on the new filename
- Perform edits on a file, rename it, then perform more edits on the
  new filename

The protocol seems to allow all of the above.  Although the
specification does not go into detail how successive edits are to be
handled, it should be safe to assume that the logic is the same as for
textDocument/didChange's contentChanges, i.e. each edit moves the
workspace to a new state, and the next edit is to be applied onto this
state.

Do this here too.