Be robust to invalid uri's returned in textDocuemnt/references responese
See https://github.com/emacs-lsp/lsp-mode/issues/1113, where ccls is returning an invalid empty
entry in the response. We've narrowed down the issue to a parsing problem with google test macros
and hope to help solve the ccls issue soon.

This updates the code to generically handle errors and explicitly catch empty uri's. The reason
for explicitly handling empty filenames from the uri is to help diagnose the error. If we don't
explicitly handle empty filenames, then a less informative warning is shown stating that
a directory cannot be read because (insert-file-contents-literally "") tries to insert the
current directory.