lsp--line-character-to-point: `character`  'nil? (#1440)
* lsp--line-character-to-point: `character`  'nil?

Fixes https://github.com/emacs-lsp/lsp-origami/issues/5, but I don't know what this might implicate in other places.
In certain, but not all, php-files, ```lsp--line-character-to-point``` is called with ```character``` passed as ```nil```, don't use numerical comparison in this case.  Simply check that character is not nil prior to the comparison

* Simplify the ```character``` check

```(not (equal nil character))``` is simply ```character```...

* Fixed as requested

works as well, tested locally.