Display the first line of MarkupContent in eldoc (#1607)
* Display the first line of MarkupContent in eldoc

Many LSP servers, like gopls, return `document/onHover` as
MarkupContent. This modifies the eldoc behavior to support MarkupContent
by displaying the first rendered line. As the majority of LSP servers
return the signature, this will display the signature as most people expect.

In the future,
https://github.com/microsoft/language-server-protocol/issues/772
should make this more reliable.

* nit: use s-lines