Cancel lenses requests per buffer. (#1776)
Currently lenses requests have common global cancel token which means
we can't send lenses requests for different buffer simultaneously.
But there are cases when we receive a notification from a server that
lenses where changed for all buffers (lsp-metals) and we need to
refresh them. Currently lenses are not refreshed after buffer switch
which also means that lenses can be lost if user switches buffers
right after a buffer change.