Horizontal rule affects lsp-ui-doc width
Fixes emacs-lsp/lsp-mode#442. Adds a few more patterns to remove before
rendering markdown. Because there are multiple ways to represent a horizontal
rule (<hr/>) in markdown, the original attempts to fix this did not hit them
all. The issue manifests in an extremely large lsp-ui-doc popup window. Because
the horizontal rule is extending the length of the window, and couldn't be
controlled by the variables available to control the max width of a child frame,
this 'hack' was required to simply strip out the offending markdown. This has
the side affect of not showing the horizontal rules at all.