Watch certain subdirectories rather than the whole project
Watching is recursive. When watching a project, it registers a watch
on every subdirectory in the project. With node modules and
compilation output, this can get enormous, easily hitting the watch
count.

Setting `lsp-watch-folders` can specify the source files to be watched
rather than the every subdirectory of the project.

`hack-local-variables` to ensure dir-locals are applied at start

without this, on first startup dir local values for
`lsp-watch-folders` are not found. They _are_ found on restart. If we
hack local variables here they are found on startup.

Correctly cleanup watched sub folders

Want to avoid watching the root but still want the watches to be
registered under the root so the current cleanup apparatus can find it
and clean up. Put a top level `lsp-watch-root-folder` which can take
subdirectories and put each of these watch subtrees under the top
level root.

Considered using a defcustom allow-list similar to the
`lsp-file-watch-ignored` but don't want to consider these at all
levels but solely under the top level directory.