Search for clangd executable in the path with names variations.
* lsp-clients.el (lsp-clients--clangd-command) : Use some logic to
search for the clangd executable in the path with different names.

(lsp-clients-clangd-executable) : Change default value to nil to search
in the path automatically. If the user sets a different name that one
will be preferred in the search.

(lsp-clients-clangd-executable-found) : Internal variable to set the
full path for the clangd executable when found.