`lsp--symbol-ignore': ensure that kind is bounded (#2180)
* `lsp--symbol-filter': ensure that kind is bounded

Servers like ccls use 'kind' values outside the normal range when
describing symbols that lie outside the protocol. We set those values
to zero and allow nil to represent the corresponding symbols in
`lsp-imenu-index-symbol-kinds'.

* Rename `lsp--symbol-filter' to `lsp--symbol-ignore'

The word 'filter' does not say whether the predicate function keeps or
tosses the elements that satisfy it.