`lsp-file-watch-ignored': use proper regexes
Emacs regexes specify the end of string using \', not $ as was used in that
variable (the latter matching *either* the end of the line or the string, which
could cause issues with files that have newlines in them (Linux allows this)).
Fix that by using \' for the end of string character instead.

Also properly handle node_modules. It is not necessarily always at the root of
the project, as is the case with .pyi files needed by pyright, which are in a
node_modules folder somewhere in the user's home directory. This causes
annoyances after Emacs-restarts, with `lsp' prompting the user to watch 11+k
immutable files. Fix that by removing the `eos' matcher and matching a slash
after "node_modules".