fix "using full path on remote host" #2251
When using a full path for a Language Server (like
`/ssh:remotehost/home/me/bin/pyls`), this will handle the remote
location correctly.

Previously, the command was passed with remote/tramp prefix, but on
the remote machine, an error would happen because
`/ssh:remotehost/home/me/bin/pyls` is not a real file that exists there.