Centralize SymbolInformation rendering
In various different parts of `lsp', where SymbolInformation has to be
rendered (lsp-ivy, lsp-treemacs, lsp-imenu), only the :name is shown as
a string. This is insufficient, as other information is then ignore
displayed (currently only :deprecated?).

Add a function, `lsp-render-symbol-information', which behaves similar
to `lsp-render-symbol'. Use it in `lsp--imenu-create-non-hierarchical'
index.