`lsp-avy-lens': don't `error' for no lenses
If there are no elements, `avy-process' returns t, causing
`funcall-interactively' to error. Before continuing to that function,
check if there are lenses, only then calling `avy'. If there are none,
`user-error' out.

Replacing the guard from the previous commit with `consp' was also an
option, but that would yield a less intuitive "zero candidates" message
instead.