`lsp-avy-lens': don't `error' for no lenses (#2339)
* `lsp-avy-lens': don't `error' for no lenses

If there are no elements, `avy-process' returns t, causing
`funcall-interactively' to error. Before continuing to that function,
check if there are lenses, only then calling `avy'. If there are none,
`user-error' out.

Replacing the guard from the previous commit with `consp' was also an
option, but that would yield a less intuitive "zero candidates" message
instead.

* `lsp-avy-lens': remove `lsp-lens-mode' check

Given that we already check for `lsp-lens--overlays', it makes no sense to check
for `lsp-lens-mode' first, because `lsp-lens--overlays' can only have lenses if
that mode is active. In addition, not checking for that mode decouples
`lsp-avy-lens' from `lsp-lens-mode'.

* Revert action -> actions

No longer use an ACTIONS variable to hold the result of `avy-process',
because that is an incorrect name (it is in fact a tuple of values, only
the third of which is an action). Instead, revert to the old state,
where ACTION was acquired using `cl-third' directly.

* Make a regex less confusing: "\|" -> "|"

"\|" = "|", because an escaped slash is just a slash, but the former
seems like a misspelled version of "\\|", possibly causing confusion.