`lsp--select-action': don't show disabled actions
In that function, filter the actions to only include non-disabled ones. This
way, `lsp-no-code-actions' is `signal'ed even when there are code actions that
are all `:disabled?'.

This is consistent with the spec guidelines: "disabled code actions are
not shown in automatic light bulb code action menus".

Given that we now support it at least somewhere, set the
"disabledSupport" capability.