`lsp-execute-code-action': C-u -> show all actions
With a C-u prefix, `lsp-execute-code-action' now also shows disabled
code actions. These are propertized with
`lsp-disabled-code-action-face', and have the reason appended to them.

To implement this, make `lsp--select-action' no longer filter the
actions, and to use the new `lsp--code-action-title' to render their
tiles. The new function checks whether the code action is preferred or
disabled and returns a proper title.

A new function, `lsp--select-enabled-action' is the same as the above
but filters out disabled candidates. Use it in
`lsp-execute-code-action-by-kind', because it is usually used to execute
specific, global actions where it wouldn't make sense to have a prompt
if they are unavailable.