Fix keyword auto-completion in classes for iph
When auto-completing public, ... at the start of a class, Intelephense
sends `CompletionList's that have no :isIncomplete fields (a spec
violation), causing `lsp-completion--sort-completions' to get a
`CompletionList' instead of a list of `CompletionItem', `error'ing.

Extend `lsp-interface' to take an optional fourth list of semi-optional
arguments (like optional but without a ?) for each interface tuple.
Define `CompletionList' so that is-optional is semi-optional.