`lsp-lens-mode': disabling now works
Now disabling `lsp-lens-mode' actually works, and the lenses don't
reappear after a short delay. The issue was that the wrong lambdas were
being removed from the various hooks in `lsp-lens-mode' and because
`lsp-lens--refresh-timer' was not cancelled.

To fix this, extract `(lsp--lens-schedule-refresh t)' to a new function
instead, since they are much less fragile in hooks. At the end of the
disable part of `lsp-lens-mode', cancel `lsp-lens--refresh-timer'.