Bugfix/lens check if supported (#2389)
* lsp-lens: only enable if supported

`lsp--auto-configure': add `lsp-lens--enable' instead of
`lsp-lens-mode', since the former checks for lens support. This way, the
mode won't be enabled for servers that don't support lenses.

* `lsp-lens-mode': disabling now works

Now disabling `lsp-lens-mode' actually works, and the lenses don't
reappear after a short delay. The issue was that the wrong lambdas were
being removed from the various hooks in `lsp-lens-mode' and because
`lsp-lens--refresh-timer' was not cancelled.

To fix this, extract `(lsp--lens-schedule-refresh t)' to a new function
instead, since they are much less fragile in hooks. At the end of the
disable part of `lsp-lens-mode', cancel `lsp-lens--refresh-timer'.

* `lsp-lens-mode': check if timer is non-nil

`lsp-lens--refresh-timer' could be nil if `lsp-lens-mode' is disabled
within `lsp-idle-delay' afterwards.

* `lsp-lens-mode': dont configure globally

`lsp-lens-mode' is not responsible for globally configuring lenses, so
it shouldn't modify `lsp-configure-hook' and `lsp-unconfigure-hook'.