`lsp-execute-code-action-by-kind': refactor
The `if-let' around the call to `lsp--select-enabled-action' is
redundant, because that function can't ever return nil: if there are no
code actions, it signals `lsp-no-code-actions' (as the old code did),
and for the user to select nothing from `completing-read', they would
have to `quit', aborting execution.

This will allow that check to be ommitted from similar functions while
retaining code consistency.