`lsp--select-action': preselect preferred actions
Extend `lsp-completing-read' to allow DEF to be a number the index into
collection that is to be pre-selected. In `lsp--select-action', pass the
index of the first `lsp:code-action-is-preferred' item in the given list
as that argument, pre-selecting it.

Using an index into the collection is needed in `lsp-completing-read',
because the given TRANSFORMER-FN is often not side-effect-free (as is
the case in `lsp--select-action', due to
`lsp--create-unique-string-fn'), so transforming whatever default is to
be used first and passing the resulting string as DEF is not possible.