Add an `iedit' command to edit documentHighlights
Add a new command, `lsp-iedit-highlights', which will invoke `iedit' on
the documentHighlights at point. It can be used as a primitive
`lsp-rename' replacement if the former is unsupported.

Implement it on top the new function, `lsp-iedit--on-ranges', which
invokes `iedit' on a list of `&Range's. This separation will allow
implementing a similar `iedit' commmand for `LinkedEditingRanges'.