`lsp-iedit': move to separate file, fix byte-comp
Move `lsp-iedit' to a new file, lsp-iedit.el. Don't do that for
`lsp--range-text' though, because it is unrelated to `iedit' and should
as such be part of the core.

`declare' and `defvar' `iedit' variables, so that the new module now
byte-compiles cleanly.

`require' `iedit', so that `lsp-iedit-highlights' will work even if the
user hasn't used `iedit-mode' first.