Add an `iedit' command to edit documentHighlights (#2477)
* Add an `iedit' command to edit documentHighlights

Add a new command, `lsp-iedit-highlights', which will invoke `iedit' on
the documentHighlights at point. It can be used as a primitive
`lsp-rename' replacement if the former is unsupported.

Implement it on top the new function, `lsp-iedit--on-ranges', which
invokes `iedit' on a list of `&Range's. This separation will allow
implementing a similar `iedit' commmand for `LinkedEditingRanges'.

* `lsp-iedit': move to separate file, fix byte-comp

Move `lsp-iedit' to a new file, lsp-iedit.el. Don't do that for
`lsp--range-text' though, because it is unrelated to `iedit' and should
as such be part of the core.

`declare' and `defvar' `iedit' variables, so that the new module now
byte-compiles cleanly.

`require' `iedit', so that `lsp-iedit-highlights' will work even if the
user hasn't used `iedit-mode' first.

* lsp-iedit.el: add Copyright header

* lsp-iedit.el: consistent declare-function, defvar

lsp-mode.el places `declare-function's near the top of the file,
followed by a `defvar' block, separated by newlines. Use the same style
in lsp-iedit.el.