Add a new categorized `imenu' index
`lsp-imenu-create-categorized-index' becomes
`lsp-imenu-create-top-level-categorized-index', with the former now now
creating a "traditional" `imenu' index where all Functions are under a
`Functions' section, all Variables under `Variables', etc. It works with
both `SymbolInformation' and `DocumentSymbol's (tested with `pyright'
and `iph', respectively).

Currently there is no way to jump to a symbol if it has children; this
defect is shared by `org-imenu', because `imenu' entries with children
cannot encode a position.